Make today's mood a new component
